geometry script
